resolver: inline small h3 module
